this is a huge commit: added -cimg+linux now working -cimg plugin that is capable of loading any order of bgra/bgr pixels to shared/non-shared images -cimg resize interpolation type combobox in controldialog for playing around with the interpolation type -capability to reach ~75 fps using cimg grid interpolation

+++ ambilight-host/src/letterboxometer.cpp
@@ -45,8 +45,8 @@ std::pair<int, int> LetterboxOMeter::measure() {
 }
 
 bool LetterboxOMeter::isBlack(Image img) {
-    for (size_t x = 0; x < img.width(); x++) {
-        for (size_t y = 0; y < img.height(); y++) {
+    for (int x = 0; x < img.width(); x++) {
+        for (int y = 0; y < img.height(); y++) {
             if (img(x, y, CIMG_2D_Z_LEVEL, CIMG_RED_CHANNEL) > mBlackThreshold) return false;
             if (img(x, y, CIMG_2D_Z_LEVEL, CIMG_GREEN_CHANNEL) > mBlackThreshold) return false;
             if (img(x, y, CIMG_2D_Z_LEVEL, CIMG_BLUE_CHANNEL) > mBlackThreshold) return false;
